[FEATURE] Add a DummyConfiguration class
This class can be used in tests.
